Core/Misc: Fixed wrong db field accessor functions use
Port From (https://github.com/TrinityCore/TrinityCore/commit/96ec1f16d434b65ac7775ff9fa50f605f6fbf4d8)

+++ Source/Game/Cache/CharacterCache.cs
@@ -29,7 +29,7 @@ namespace Game.Cache
 
             do
             {
-                AddCharacterCacheEntry(ObjectGuid.Create(HighGuid.Player, result.Read<uint>(0)), result.Read<uint>(2), result.Read<string>(1), result.Read<byte>(4), result.Read<byte>(3), result.Read<byte>(5), result.Read<byte>(6), result.Read<uint>(7) != 0);
+                AddCharacterCacheEntry(ObjectGuid.Create(HighGuid.Player, result.Read<ulong>(0)), result.Read<uint>(2), result.Read<string>(1), result.Read<byte>(4), result.Read<byte>(3), result.Read<byte>(5), result.Read<byte>(6), result.Read<uint>(7) != 0);
             } while (result.NextRow());
 
             Log.outInfo(LogFilter.ServerLoading, $"Loaded character infos for {_characterCacheStore.Count} characters in {Time.GetMSTimeDiffToNow(oldMSTime)} ms");
